Add files via upload
While there are examples in ExampleMod of modifying the world generation process itself, such as ExampleOre, there are no examples of adding to or adjusting any of the Hardmode generation tasks. This PR proposes to add a simple system file to the ExampleMod that demonstrates how to add a basic task to the Hardmode task list which could then be extrapolated upon by learning modders.